Fix adress length of CAN bus and add it to peer information

+++ components/robusto/include/robusto_peer.h
@@ -194,7 +194,7 @@ extern "C"
      * @param canbus_address The CAN bus address we are looking for
      * @return robusto_peer_t* 
      */
-    robusto_peer_t *robusto_peers_find_peer_by_canbus_address(uint8_t canbus_address);
+    robusto_peer_t *robusto_peers_find_peer_by_canbus_address(uint32_t canbus_address);
 #endif 
 
 
@@ -252,7 +252,7 @@ extern "C"
         uint8_t i2c_address;
 #endif
 #ifdef CONFIG_ROBUSTO_SUPPORTS_CANBUS
-        uint8_t canbus_address;
+        uint32_t canbus_address;
 #endif
     } relation_t;
 
@@ -295,7 +295,7 @@ extern "C"
 #endif
 #ifdef CONFIG_ROBUSTO_SUPPORTS_CANBUS
                       ,
-                      uint8_t canbus_address
+                      uint32_t canbus_address
 #endif
     );
